fix new linked section (#1108)

+++ modules/Nncase.Modules.K210/CodeGen/K210/KPULinkedModule.cs
@@ -17,8 +17,8 @@ internal class KPULinkedModule : ILinkedModule
         Functions = functions;
         Sections = new[]
         {
-            new LinkedSection(text, ".text", 0, 8, (uint)text.Length),
-            new LinkedSection(rdata, ".rdata", 0, 8, (uint)(rdata?.Length ?? 0)),
+            new LinkedSection(text, ".text", 0, 8, (ulong)text.Length),
+            new LinkedSection(rdata, ".rdata", 0, 8, (ulong)(rdata?.Length ?? 0)),
         };
     }
