+++ modules/Nncase.Modules.StackVM/CodeGen/StackVM/StackVMLinkedModule.cs
@@ -17,9 +17,9 @@ internal class StackVMLinkedModule : ILinkedModule
         Functions = functions;
         Sections = new[]
         {
-            new LinkedSection(text, ".text", 0, 8, (uint)text.Length),
-            new LinkedSection(rdata, ".rdata", 0, 8, (uint)(rdata?.Length ?? 0)),
-            new LinkedSection(custom_calls, ".custom_calls", 0, 8, (uint)(custom_calls?.Length ?? 0)),
+            new LinkedSection(text, ".text", 0, 8, (ulong)text.Length),
+            new LinkedSection(rdata, ".rdata", 0, 8, (ulong)(rdata?.Length ?? 0)),
+            new LinkedSection(custom_calls, ".custom_calls", 0, 8, (ulong)(custom_calls?.Length ?? 0)),
         };
     }
